Move getBlockNumber to common utility file
